Isolate site-specific discovery adapters
