+++ dezoomify-core/src/zoomify/mod.rs
@@ -17,7 +17,7 @@ mod image_properties;
 const ROUTES: &[DiscoveryRoute] = &[
     DiscoveryMatch::UrlPredicate(is_tile_url).map_url(tile_metadata),
     DiscoveryMatch::UrlSuffix("ImageProperties.xml").extract(load_catalog),
-    DiscoveryMatch::ContentPredicate(contains_ngv_image_path).then(follow_ngv_image_path),
+    DiscoveryMatch::ContentPredicate(ngv::contains_image_path).then(ngv::follow_image_path),
     DiscoveryMatch::ContentPredicate(contains_zoomify_declaration)
         .then(extract_image_properties_url),
 ];
@@ -47,11 +47,6 @@ static HTML_BASE_RE: LazyLock<BytesRegex> = LazyLock::new(|| {
     BytesRegex::new(r#"(?is)<base\s+[^>]*\bhref\s*=\s*["'](?P<base>[^"']*)"#)
         .expect("constant HTML base pattern")
 });
-static NGV_IMAGE_PATH_RE: LazyLock<BytesRegex> = LazyLock::new(|| {
-    BytesRegex::new(r#"(?is)\bvar\s+url\s*=\s*['"](?P<image>[^'"]+)['"]"#)
-        .expect("constant NGV Zoomify path pattern")
-});
-
 static TILE_URL_RE: LazyLock<Regex> = LazyLock::new(|| {
     Regex::new(r"(?:^|/)TileGroup\d+/\d+-\d+-\d+\.jpe?g(?:[?#].*)?$")
         .expect("constant Zoomify tile URL pattern")
@@ -62,30 +57,7 @@ fn is_tile_url(uri: &str) -> bool {
 }
 
 fn is_zoomify_url(uri: &str) -> bool {
-    uri.contains("/ImageProperties.xml")
-        || uri.contains("ngv.vic.gov.au/explore/collection/work")
-        || is_tile_url(uri)
-}
-
-fn contains_ngv_image_path(contents: &[u8]) -> bool {
-    NGV_IMAGE_PATH_RE.is_match(contents)
-}
-
-fn follow_ngv_image_path(
-    _: &DiscoveryContext<'_>,
-    resource: crate::core::DiscoveryResource<'_>,
-) -> Result<DiscoveryStep, DiscoveryError> {
-    let path = NGV_IMAGE_PATH_RE
-        .captures(resource.bytes())
-        .and_then(|captures| capture_text(&captures, "image"))
-        .ok_or_else(|| {
-            DiscoveryError::Session("NGV page does not declare a Zoomify path".into())
-        })?;
-    let image_uri = resolve_relative(resource.final_uri(), &path);
-    Ok(DiscoveryStep::Follow(Request::new(append_path_component(
-        &image_uri,
-        "ImageProperties.xml",
-    ))))
+    uri.contains("/ImageProperties.xml") || ngv::is_work_page(uri) || is_tile_url(uri)
 }
 
 fn extract_image_properties_url(
@@ -105,6 +77,8 @@ fn extract_image_properties_url(
         "ImageProperties.xml",
     ))))
 }
+
+mod ngv;
 
 fn contains_zoomify_declaration(contents: &[u8]) -> bool {
     SHOW_IMAGE_RE.is_match(contents)
